Fixing maximum power charging strategy
